+++ _install.php
@@ -3,7 +3,7 @@
 #
 # This file is part of Related, a plugin for DotClear2.
 #
-# Copyright(c) 2014 Nicolas Roudaire <nikrou77@gmail.com> http://www.nikrou.net
+# Copyright(c) 2014-2015 Nicolas Roudaire <nikrou77@gmail.com> http://www.nikrou.net
 #
 # Copyright (c) 2006-2010 Pep and contributors.
 # Licensed under the GPL version 2.0 license.
@@ -11,6 +11,7 @@
 # http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
 #
 # -- END LICENSE BLOCK ------------------------------------
+
 if (!defined('DC_CONTEXT_ADMIN')) return;
 
 $this_version = $core->plugins->moduleInfo('related','version');
@@ -20,7 +21,9 @@ if (version_compare($installed_version,$this_version,'>=')) {
 }
 
 $core->blog->settings->addNamespace('related');
-if (!$core->blog->settings->related->related_files_path) {
+$core->blog->settings->related->put('active', false, 'boolean', 'Related plugin activated?', false);
+
+if (!$core->blog->settings->related->files_path) {
 	$public_path = $core->blog->public_path;
 	$related_files_path = $public_path.'/related';
 
@@ -28,12 +31,10 @@ if (!$core->blog->settings->related->related_files_path) {
 		if (!is_readable($related_files_path) || !is_writable($related_files_path)) {
 			throw new Exception(__('Directory for related files repository needs to allow read and write access.'));
 		}
-	}
-	else {
+	} else {
 		try {
 			files::makeDir($related_files_path);
-		}
-		catch (Exception $e) {
+		} catch (Exception $e) {
 			throw $e;
 		}
 	}
@@ -41,12 +42,11 @@ if (!$core->blog->settings->related->related_files_path) {
 	if (!is_file($related_files_path.'/.htaccess')) {
 		try {
 			file_put_contents($related_files_path.'/.htaccess',"Deny from all\n");
-		}
-		catch (Exception $e) {}
+		} catch (Exception $e) {}
 	}
 
-	$core->blog->settings->related->put('related_url_prefix','static', 'string', 'Prefix used by the URLHandler',true);
-	$core->blog->settings->related->put('related_files_path',$related_files_path, 'string', 'Related files repository',true);
+	$core->blog->settings->related->put('url_prefix','static', 'string', 'Prefix used by the URLHandler',true);
+	$core->blog->settings->related->put('files_path',$related_files_path, 'string', 'Related files repository',true);
 }
 
 $core->setVersion('related',$this_version);