+++ index.php
@@ -3,7 +3,7 @@
 #
 # This file is part of Related, a plugin for DotClear2.
 #
-# Copyright(c) 2014 Nicolas Roudaire <nikrou77@gmail.com> http://www.nikrou.net
+# Copyright(c) 2014-2015 Nicolas Roudaire <nikrou77@gmail.com> http://www.nikrou.net
 #
 # Copyright (c) 2006-2010 Pep and contributors.
 # Licensed under the GPL version 2.0 license.
@@ -11,12 +11,11 @@
 # http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
 #
 # -- END LICENSE BLOCK ------------------------------------
+
 if (!defined('DC_CONTEXT_ADMIN')) return;
 
-$what = (!empty($_REQUEST['do']) && $_REQUEST['do'] == 'edit') ? 'page' : 'panel';
-if ($what == 'page') {
-	require_once dirname(__FILE__).'/inc/page.php';
-}
-else {
-	require_once dirname(__FILE__).'/inc/panel.php';
+if (!empty($_REQUEST['do']) && $_REQUEST['do'] == 'edit') {
+    include_once dirname(__FILE__).'/inc/page.php';
+} else {
+    include_once dirname(__FILE__).'/inc/panel.php';
 }