+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+related 1.2.0 - 2015-01-02
+==========================
+* Introduce related documentation
+* Plugin can be activated per blog
+
 related 1.1.1 - 2014-12-31
 ==========================
 * problem with attachment. Remove behavior for now.